+++ pythonFiles/include/blender_vscode/__init__.py
@@ -1,5 +1,6 @@
 import sys
 from dataclasses import dataclass
+from typing import List
 from pathlib import Path
 from typing import List