+++ pythonFiles/include/blender_vscode/load_addons.py
@@ -1,48 +1,196 @@
 import os
+import subprocess
 import sys
 import traceback
 from pathlib import Path
-from typing import List
+from typing import List, Union, Optional, Dict
 
 import bpy
 
 from . import AddonInfo
 from .communication import send_dict_as_json
 from .environment import addon_directories
-from .utils import is_addon_legacy
+from .utils import is_addon_legacy, addon_has_bl_info
+
+if bpy.app.version >= (4, 2, 0):
+    _EXTENSIONS_DEFAULT_DIR = Path(bpy.utils.user_resource("EXTENSIONS", path="user_default"))
+else:
+    _EXTENSIONS_DEFAULT_DIR = None
+_ADDONS_DEFAULT_DIR = Path(bpy.utils.user_resource("SCRIPTS", path="addons"))
 
 
-def setup_addon_links(addons_to_load: List[AddonInfo]):
+def setup_addon_links(addons_to_load: List[AddonInfo]) -> List[Dict]:
+    path_mappings: List[Dict] = []
 
-    path_mappings = []
+    # always make sure addons are in path, important when running fresh blender install
+    # do it always to avoid very confusing logic in the loop below
+    os.makedirs(_ADDONS_DEFAULT_DIR, exist_ok=True)
+    if str(_ADDONS_DEFAULT_DIR) not in sys.path:
+        sys.path.append(str(_ADDONS_DEFAULT_DIR))
+
+    remove_broken_addon_links()
+    if bpy.app.version >= (4, 2, 0):
+        remove_broken_extension_links()
 
     for addon_info in addons_to_load:
-        user_addon_directory = get_user_addon_directory(Path(addon_info.load_dir))
-        print(f"USER ADDON: {user_addon_directory}")
-
-        if not os.path.exists(user_addon_directory):
-            os.makedirs(user_addon_directory)
-
-        if is_addon_legacy(Path(addon_info.load_dir)) and not str(user_addon_directory) in sys.path:
-            sys.path.append(str(user_addon_directory))
-
-        if is_in_any_addon_directory(addon_info.load_dir):
-            load_path = addon_info.load_dir
+        try:
+            load_path = _link_addon_or_extension(addon_info)
+        except PermissionError as e:
+            print(
+                f"""ERROR: {e} 
+Path "{e.filename}" can not be removed. **Please remove it manually!** Most likely causes:
+    - Path requires admin permissions to remove
+    - Windows only: You upgraded Blender version and imported old setting. Now links became real directories.
+    - Path is a real directory with the same name as addon (removing might cause data loss!)"""
+            )
+            raise e
         else:
-            load_path = os.path.join(user_addon_directory, addon_info.module_name)
-            create_link_in_user_addon_directory(addon_info.load_dir, load_path)
-
-        path_mappings.append({"src": str(addon_info.load_dir), "load": str(load_path)})
+            path_mappings.append({"src": str(addon_info.load_dir), "load": str(load_path)})
 
     return path_mappings
 
 
-def get_user_addon_directory(source_path: Path):
-    """Return either the user scripts or user extensions directory depending on the addon type."""
-    if is_addon_legacy(source_path):
-        return Path(bpy.utils.user_resource("SCRIPTS", path="addons"))
+def _link_addon_or_extension(addon_info: AddonInfo) -> Path:
+    if is_addon_legacy(addon_info.load_dir):
+        if is_in_any_addon_directory(addon_info.load_dir):
+            # blender knows about addon and can load it
+            load_path = addon_info.load_dir
+        else:  # addon is in external dir or is in extensions dir
+            _remove_duplicate_addon_links(addon_info)
+            load_path = _ADDONS_DEFAULT_DIR / addon_info.module_name
+            create_link_in_user_addon_directory(addon_info.load_dir, load_path)
     else:
-        return Path(bpy.utils.user_resource("EXTENSIONS", path="user_default"))
+        if addon_has_bl_info(addon_info.load_dir) and is_in_any_addon_directory(addon_info.load_dir):
+            # this addon is compatible with legacy addons and extensions
+            # but user is developing it in addon directory. Treat it as addon.
+            load_path = addon_info.load_dir
+        elif is_in_any_extension_directory(addon_info.load_dir):
+            # blender knows about extension and can load it
+            load_path = addon_info.load_dir
+        else:
+            # blender does not know about extension, and it must be linked to default location
+            _remove_duplicate_extension_links(addon_info)
+            os.makedirs(_EXTENSIONS_DEFAULT_DIR, exist_ok=True)
+            load_path = _EXTENSIONS_DEFAULT_DIR / addon_info.module_name
+            create_link_in_user_addon_directory(addon_info.load_dir, load_path)
+    return load_path
+
+
+def _remove_duplicate_addon_links(addon_info: AddonInfo):
+    existing_addon_with_the_same_target = does_addon_link_exist(addon_info.load_dir)
+    while existing_addon_with_the_same_target:
+        if existing_addon_with_the_same_target:
+            print("INFO: Removing old link:", existing_addon_with_the_same_target)
+            os.remove(existing_addon_with_the_same_target)
+        existing_addon_with_the_same_target = does_addon_link_exist(addon_info.load_dir)
+
+
+def _remove_duplicate_extension_links(addon_info: AddonInfo):
+    existing_extension_with_the_same_target = does_extension_link_exist(addon_info.load_dir)
+    while existing_extension_with_the_same_target:
+        if existing_extension_with_the_same_target:
+            print("INFO: Removing old link:", existing_extension_with_the_same_target)
+            os.remove(existing_extension_with_the_same_target)
+        existing_extension_with_the_same_target = does_extension_link_exist(addon_info.load_dir)
+
+
+def _resolve_link_windows_cmd(path: Path) -> Optional[str]:
+    IO_REPARSE_TAG_MOUNT_POINT = "0xa0000003"
+    JUNCTION_INDICATOR = f"Reparse Tag Value : {IO_REPARSE_TAG_MOUNT_POINT}"
+    try:
+        output = subprocess.check_output(["fsutil", "reparsepoint", "query", str(path)])
+    except subprocess.CalledProcessError:
+        return None
+    output_lines = output.decode().split(os.linesep)
+    if not output_lines[0].startswith(JUNCTION_INDICATOR):
+        return None
+    TARGET_PATH_INDICATOR = "Print Name:            "
+    for line in output_lines:
+        if line.startswith(TARGET_PATH_INDICATOR):
+            possible_target = line[len(TARGET_PATH_INDICATOR) :]
+            if os.path.exists(possible_target):
+                return possible_target
+
+
+def _resolve_link(path: Path) -> Optional[str]:
+    """Return target if is symlink or junction"""
+    try:
+        return os.readlink(str(path))
+    except OSError as e:
+        # OSError: [WinError 4390] The file or directory is not a reparse point
+        if sys.platform == "win32":
+            if e.winerror == 4390:
+                return None
+        else:
+            # OSError: [Errno 22] Invalid argument: '/snap/blender/5088/4.2/extensions/system/readme.txt'
+            if e.errno == 22:
+                return None
+        print("Warning: can not resolve link target", e)
+        return None
+    except ValueError as e:
+        # there are major differences in python windows junction support (3.7.0 and 3.7.9 give different errors)
+        if sys.platform == "win32":
+            return _resolve_link_windows_cmd(path)
+        else:
+            print("Warning: can not resolve link target", e)
+            return None
+
+
+def does_addon_link_exist(development_directory: Path) -> Optional[Path]:
+    """Search default addon path and return first path that links to `development_directory`"""
+    for file in os.listdir(_ADDONS_DEFAULT_DIR):
+        existing_addon_dir = Path(_ADDONS_DEFAULT_DIR, file)
+        target = _resolve_link(existing_addon_dir)
+        if target:
+            windows_being_windows = target.lstrip(r"\\?")
+            if Path(windows_being_windows) == Path(development_directory):
+                return existing_addon_dir
+    return None
+
+
+def does_extension_link_exist(development_directory: Path) -> Optional[Path]:
+    """Search all available extension paths and return path that links to `development_directory"""
+    for repo in bpy.context.preferences.extensions.repos:
+        if not repo.enabled:
+            continue
+        repo_dir = repo.custom_directory if repo.use_custom_directory else repo.directory
+        if not os.path.isdir(repo_dir):
+            continue  # repo dir might not exist
+        for file in os.listdir(repo_dir):
+            existing_extension_dir = Path(repo_dir, file)
+            target = _resolve_link(existing_extension_dir)
+            if target:
+                windows_being_windows = target.lstrip(r"\\?")
+                if Path(windows_being_windows) == Path(development_directory):
+                    return existing_extension_dir
+    return None
+
+
+def remove_broken_addon_links():
+    for file in os.listdir(_ADDONS_DEFAULT_DIR):
+        addon_dir = _ADDONS_DEFAULT_DIR / file
+        if not addon_dir.is_dir():
+            continue
+        target = _resolve_link(addon_dir)
+        if target and not os.path.exists(target):
+            print("INFO: Removing invalid link:", addon_dir, "->", target)
+            os.remove(addon_dir)
+
+
+def remove_broken_extension_links():
+    for repo in bpy.context.preferences.extensions.repos:
+        if not repo.enabled:
+            continue
+        repo_dir = repo.custom_directory if repo.use_custom_directory else repo.directory
+        repo_dir = Path(repo_dir)
+        if not repo_dir.is_dir():
+            continue
+        for file in os.listdir(repo_dir):
+            existing_extension_dir = repo_dir / file
+            target = _resolve_link(existing_extension_dir)
+            if target and not os.path.exists(target):
+                print("INFO: Removing invalid link:", existing_extension_dir, "->", target)
+                os.remove(existing_extension_dir)
 
 
 def load(addons_to_load: List[AddonInfo]):
@@ -50,10 +198,16 @@ def load(addons_to_load: List[AddonInfo]):
         if is_addon_legacy(Path(addon_info.load_dir)):
             bpy.ops.preferences.addon_refresh()
             addon_name = addon_info.module_name
+        elif addon_has_bl_info(addon_info.load_dir) and is_in_any_addon_directory(addon_info.load_dir):
+            # this addon is compatible with legacy addons and extensions
+            # but user is developing it in addon directory. Treat it as addon.
+            bpy.ops.preferences.addon_refresh()
+            addon_name = addon_info.module_name
         else:
             bpy.ops.extensions.repo_refresh_all()
-            addon_name = "bl_ext.user_default." + addon_info.module_name
-
+            repo = is_in_any_extension_directory(addon_info.load_dir)
+            module = getattr(repo, "module", "user_default")
+            addon_name = ".".join(("bl_ext", module, addon_info.module_name))
         try:
             bpy.ops.preferences.addon_enable(module=addon_name)
         except Exception:
@@ -61,7 +215,7 @@ def load(addons_to_load: List[AddonInfo]):
             send_dict_as_json({"type": "enableFailure", "addonPath": str(addon_info.load_dir)})
 
 
-def create_link_in_user_addon_directory(directory, link_path):
+def create_link_in_user_addon_directory(directory: Union[str, os.PathLike], link_path: Union[str, os.PathLike]):
     if os.path.exists(link_path):
         os.remove(link_path)
 
@@ -73,8 +227,18 @@ def create_link_in_user_addon_directory(directory, link_path):
         os.symlink(str(directory), str(link_path), target_is_directory=True)
 
 
-def is_in_any_addon_directory(module_path):
+def is_in_any_addon_directory(module_path: Path) -> bool:
     for path in addon_directories:
         if path == module_path.parent:
             return True
     return False
+
+
+def is_in_any_extension_directory(module_path: Path) -> Optional["bpy.types.UserExtensionRepo"]:
+    for repo in bpy.context.preferences.extensions.repos:
+        if not repo.enabled:
+            continue
+        repo_dir = repo.custom_directory if repo.use_custom_directory else repo.directory
+        if Path(repo_dir) == module_path.parent:
+            return repo
+    return None